+++ NEWS.md
@@ -1,5 +1,8 @@
 # rvest (development version)
 
+* `submit_form()` now allows you to pick the submission button by position
+  (#156).
+
 * `set_values()` now uses dynamic dots so that you can use `!!!` if you have
   a list of values (#189).